+++ app/View/users/delete-users.php
@@ -1,19 +1,19 @@
 <?php
-include_once '../app/Model/connection-pdo.php';
+if (isset($_POST['idUser']) && !empty($_POST['idUser'])) {
+    
+    include_once '../app/Model/connection-pdo.php';
 
-$idUser = filter_input(INPUT_POST, 'idUser', FILTER_SANITIZE_NUMBER_INT);
-
-if (!empty($idUser)) {
+    $idUser = filter_input(INPUT_POST, 'idUser', FILTER_SANITIZE_NUMBER_INT);
     $queryDeleteUser = " DELETE FROM tb_usuario WHERE usu_codigo=:usu_codigo ";
     $deleteUser = $connectionDataBase->prepare($queryDeleteUser);
-
     $deleteUser->bindParam("usu_codigo", $idUser);
-    $deleteUser->execute();
 
-    $_SESSION['msg'] = 'Inserido com sucesso';
-    header('Location: Pagina.php');
-
-} else {
-    //header("Location: /agenda/calendario");
-} 
+    if ($deleteUser->execute()) {
+        $returnAjax = true;
+    }else {
+        $returnAjax = false;
+    }
+    header('Content-Type: application/json');
+    echo json_encode($returnAjax);
+}
 ?>
